+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Support for Android 11 (March, 2021)
 
 ### Changed
+* Replaced `DecoratorCommandStrategy` with `DelegatingCommandStrategy`.
 * Replaced `IViewModelView` with `IDispatcher`.
   * Replaced `IViewModel.View` with `IViewModel.Dispatcher`
 * Update UWP target to 19041 and net framework to 4.7.2
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * [#39] Change target Uno.UI version to 4.0.7.
 
 ### Deprecated
+* `DecoratorCommandStrategy` not longer exists. Use `DelegatingCommandStrategy` instead.
 * `IViewModelView` not longer exists. Use `IDispatcher` instead.
 ### Removed
 * Support for Android 10 (March, 2022)